+++ variables.tf
@@ -1,24 +1,32 @@
 variable "feature_set" {
+  description = "The feature set of the organization. One of 'ALL' or 'CONSOLIDATED_BILLING'."
   type        = string
   default     = "ALL"
   nullable    = false
-  description = "The feature set of the organization. One of 'ALL' or 'CONSOLIDATED_BILLING'."
 }
 
 variable "aws_service_access_principals" {
+  description = "A list of AWS service principal names for which you want to enable integration with your organization."
   type        = list(string)
   default     = []
   nullable    = false
-  description = "A list of AWS service principal names for which you want to enable integration with your organization."
+}
+
+variable "enabled_policy_types" {
+  description = "The list of Organizations policy types to enable in the Organization Root. The organization must have `feature_set` set to `\"ALL\"`."
+  type        = list(string)
+  default     = []
+  nullable    = false
 }
 
 variable "organization" {
+  description = "The organization with the tree of organizational units and accounts to construct. Defaults to an object with an empty list of units and accounts"
   type = object({
     accounts = optional(list(object({
       name                              = string,
       key                               = string,
       email                             = string,
-      allow_iam_users_access_to_billing = bool,
+      allow_iam_users_access_to_billing = optional(bool, true),
     })), [])
     units = optional(list(object({
       name     = string,
@@ -27,7 +35,7 @@ variable "organization" {
         name                              = string,
         key                               = string,
         email                             = string,
-        allow_iam_users_access_to_billing = bool,
+        allow_iam_users_access_to_billing = optional(bool, true),
       })), [])
       units = optional(list(object({
         name     = string,
@@ -36,7 +44,7 @@ variable "organization" {
           name                              = string,
           key                               = string,
           email                             = string,
-          allow_iam_users_access_to_billing = bool,
+          allow_iam_users_access_to_billing = optional(bool, true),
         })), [])
         units = optional(list(object({
           name     = string,
@@ -45,7 +53,7 @@ variable "organization" {
             name                              = string,
             key                               = string,
             email                             = string,
-            allow_iam_users_access_to_billing = bool,
+            allow_iam_users_access_to_billing = optional(bool, true),
           })), [])
           units = optional(list(object({
             name     = string,
@@ -54,7 +62,7 @@ variable "organization" {
               name                              = string,
               key                               = string,
               email                             = string,
-              allow_iam_users_access_to_billing = bool,
+              allow_iam_users_access_to_billing = optional(bool, true),
             })), [])
             units = optional(list(object({
               name     = string,
@@ -63,7 +71,7 @@ variable "organization" {
                 name                              = string,
                 key                               = string,
                 email                             = string,
-                allow_iam_users_access_to_billing = bool,
+                allow_iam_users_access_to_billing = optional(bool, true),
               })), [])
             })), [])
           })), [])
@@ -73,5 +81,4 @@ variable "organization" {
   })
   default = {}
   nullable = false
-  description = "The organization with the tree of organizational units and accounts to construct. Defaults to an object with an empty list of units and accounts"
 }